requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0
flask-cors==4.0.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9